Task3_KnowledgeUpdater/requirements.txt
streamlit
wikipedia
sentence-transformers
chromadb
google-generativeai
python-dotenv